Fix bounding boxes (lat and lng inverted)
From what I gathered X and Y are currently inverted in DB.
To fix it, I changed the import method to keep the order for CRS:84
and invert it for EPSG:4326.
I also switch the fitBounds method in front to use correct LatLng order.

Breaking change : X and Y for CRS:84 needs to be inverted in DB